requirements.txt
python-telegram-bot==21.6
python-dotenv==1.2.1
fastapi
uvicorn
requests==2.32.5
